Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/referrer-policy/generic/subresource-test/script-messaging-expected.txt

Script messaging - cross-origin Script request

If you can read JSON encoded HTTP request headers of the Script below, the messaging works as expected.

http://127.0.0.1:8800/common/security-features/subresource/script.py:

{
  "accept-language": "fr-fr",
  "accept-encoding": "gzip, deflate",
  "connection": "keep-alive",
  "accept": "*/*",
  "user-agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_14_5) AppleWebKit/605.1.15 (KHTML, like Gecko)",
  "host": "127.0.0.1:8800",
  "referer": "http://localhost:8800/referrer-policy/generic/subresource-test/script-messaging.html"
}


PASS Script is responding with HTTP headers
